requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-huggingface
faiss-cpu
beautifulsoup4
python-dotenv
tiktoken